requirements.txt
Flask
Flask-Mail
python-dotenv
notion-client
openai
requests
Werkzeug
httpx
Mailtrap